move some features from eval_backend_diff.py into timing.py
